feat(Contollers):created controllers
- created a controllers Folder

- moved all controller functions to the controllers
